third_party/WebKit/LayoutTests/platform/mac-mac10.10/virtual/outofblink-cors-http/http/tests/fetch/serviceworker-proxied/thorough/auth-other-https-expected.txt

This is a testharness.js-based test.
PASS Startup 
FAIL Verify access control of fetch() in a Service Worker assert_equals: https://localhost:8443/serviceworker/resources/fetch-access-control.php?Auth&mode=cors&credentials=omit&ACAOrigin=* should be resolved expected "resolved" but got "rejected"
Harness: the test ran to completion.
